+++ l10n/es_SV.js
@@ -20,4 +20,4 @@ OC.L10N.register(
     "Tools" : "Herramientas",
     "Last modified" : "Última modificación"
 },
-"nplurals=2; plural=(n != 1);");
+"nplurals=2; plural=n == 1 ? 0 : n != 0 && n % 1000000 == 0 ? 1 : 2;");